feat: add icon help to integration fields
